Build: turn on -Wconversion -Wsign-conversion
-Wconversion -Wsign-conversion is turned on only for our library as
Paludis cannot get compiled with them. Code is fixed to respect these
flags.

+++ src/builtins/let_builtin.cpp
@@ -23,6 +23,7 @@
 #include <sstream>
 
 #include <boost/algorithm/string/join.hpp>
+#include <boost/numeric/conversion/cast.hpp>
 
 #include "core/bash_ast.h"
 
@@ -34,5 +35,5 @@ int let_builtin::exec(const std::vector<std::string>& bash_args)
   bash_ast ast(std::stringstream(expression), &bash_ast::parser_arithmetics);
   ast.interpret_with(_walker, &bash_ast::walker_arithmetics);
 
-  return ast.get_error_count();
+  return boost::numeric_cast<int>(ast.get_error_count());
 }
